+++ Semantic-Interpreter/Core/Items/Values/StringValue.cs
@@ -8,7 +8,7 @@ namespace Semantic_Interpreter.Core
         
         private string Value { get; }
 
-        public int AsInteger() => Convert.ToInt32(Value);
+        public long AsInteger() => Convert.ToInt32(Value);
 
         public double AsReal() => Convert.ToDouble(Value);
